packages/neocortex/PAR/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.28.1
anyio==4.3.0
arxiv==2.1.0
asgiref==3.8.1
asknews==0.6.8
asttokens==2.4.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
Authlib==1.3.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
dataclasses-json==0.6.4
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
elevenlabs==1.2.2
exceptiongroup==1.2.0
executing==2.0.1
feedparser==6.0.10
filelock==3.13.3
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
h11==0.14.0
html2text==2024.2.26
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.22.2
idna==3.6
ipython==8.24.0
jedi==0.19.1
jiter==0.4.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.5
langchain-anthropic==0.1.15
langchain-community==0.2.5
langchain-core==0.2.9
langchain-experimental==0.0.61
langchain-openai==0.1.8
langchain-pinecone==0.1.1
langchain-text-splitters==0.2.0
langchainhub==0.1.15
langgraph==0.0.69
langsmith==0.1.77
marshmallow==3.21.1
matplotlib-inline==0.1.7
mermaid-python==0.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.29.0
orjson==3.10.0
packaging==23.2
parso==0.8.4
pexpect==4.9.0
pinecone-client==3.2.2
playwright==1.44.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pyee==11.1.0
Pygments==2.18.0
pymongo==4.6.3
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
python-dotenv==1.0.1
pytube==15.0.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
sgmllib3k==1.0.0
six==1.16.0
slack_sdk==3.29.0
sniffio==1.3.1
soupsieve==2.5
spider-client==0.0.22
SQLAlchemy==2.0.29
stack-data==0.6.3
StackAPI==0.3.0
tenacity==8.2.3
tiktoken==0.7.0
tokenizers==0.15.2
tqdm==4.66.2
traitlets==5.14.3
types-requests==2.31.0.20240402
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
uuid6==2024.1.12
wcwidth==0.2.13
websockets==12.0
wikipedia==1.4.0
yarl==1.9.4
youtube-search==2.1.2
youtube-transcript-api==0.6.2
